Add Cycle Phases permission definitions
Ideally the Aconfig helper should be in a separated CL. But it's here :)

This doesn't actually introduce the permissions to HC (the manifest).

Note that the feature flag is in staging now.
This means adding permission definition in the manifest will make cycle
phases data type part of _every_ tests.

The record definitions are not existed yet, there's no way to make those
tests pass without making this CL gigantic. (Except for turning off the
flag)
It'll be added after the major skeleton is done.

Bug: 450576343
Flag: com.android.healthfitness.flags.cycle_phases_feature
Test: atest
Change-Id: Ifeb5cb9681e1d27318469fda7e11d60f67588646