Create a parcelable wrapper for the Device class.
It is not possible to flag a class implementing a new interface. This
was initially done in ag/36519056 and gets picked up as a new class and
after speaking with amhk@, unfortunately the presubmits and tooling does
pick up this error and doesn't support this use case. Normally it would
have been fine to add the flag to the entire existing `Device` class
go/android-sdk-docs-how-does-flaggedapi-work "disabled: the build will
advertise the API as it looks in the latest API snapshot in
prebuilts/sdk" so the API in production would have been exposed but this
isn't supported go/android-api-flagging-faq#what-cannot-be-flagged.

https://storage.googleapis.com/android-build/builds/git_aml-2026-01-release-linux-mainline_modules_sdks-user/14378150/447274ce1491d1c847be7f38470a43b31e31a30e128872fb1af43b93f865d6ea/1/mainline-sdks/for-next-build/current/com.google.android.healthfitness/sdk/healthfitness-module-sdk-current-api-diff.txt?GoogleAccessId=gcs-sign@android-builds-project.google.com.iam.gserviceaccount.com&Expires=1762340488&Signature=iBq4wg2w1zAVZZYZfQ6LDrWY91XK9rdspCcUal6q6lOngp7%2BFDWZRfFdcSihXmNtj6Vlsj4%2BGFRTh1jtjdA2dpciaE6TJ2rmfA8EOm6sb1p8IR1RlOduxcRJ5nVSgTc8RA5ts3z7QyldSsg%2BeyrGu8VYozN/2OQ2Xft4/daR5hh1Ydlz1df257FwCeUpXHkbMzdrBlWLJpQcYqNBpVoiqP6ufbdJsRUppTTCKcfD8IBfsrvkaOpo7mS5unjIVYB8dAuWipiiBkGgF%2Bcu4/Mi94cClMsf%2B74laVSl3Lrf4bUJTLm9A%2BOhD573UGdmWXEPS9qrxut%2BmLv%2BJhqfbuCyPg%3D%3D

Bug: 440055324
Bug: 416236349
Flag: com.android.healthfitness.flags.device_data_providers_api
Test: unit tests
Change-Id: I89ae0d1959208660a8f05892565868b7f472dfd5