Fix broken DDP insertion test
The broken test was first attempting to advertise the device before
verifying that when the DDP flags are off, the insertion fails with an
unsupported operation exception. Since the advertisement API would also
be disabled, the test was failing at that point. To fix this, we just
skip the advertisement.

Fix: 458044423
Test: atest HealthFitnessUnitTests com.android.server.healthconnect.HealthConnectServiceImplTest#ddpApisDisabled_insertDeviceRecords_throwsException
Flag: EXEMPT TEST_ONLY
Change-Id: I7c4a9f582d55646488ac589b34ae84290215110f