Fix advertisements, rename `device_data_provider_table` to `device_data_sources_table`.
Before, if the app_info_table contained the SPN, we did not update the table with the new app info/advertisement. We now update the app_info_table, even if an entry already exists for the SPN.

The `device_data_provider_table` has been renamed to `device_data_sources_table` to better represent what it stores. The existing DeviceDataSourcesHelper with methods for the current device has been renamed to DeviceDataSourceHelper to accommodate this.

In order to make it slightly easier to derive the SPN from a `deviceId`, the  `device_data_sources_table` now uses the `appInfoId` as a key instead of the `deviceInfoId`. A new `device_info_table` entry is made for every unique `DeviceInfo` (e.g. if the display name changes). A SPN is made up by a device type and unique DDP provided `deviceId. This `deviceId` cannot be reused for different device types (and will be enforced by b/459404842). This prevents the issue of multiple SPNs being available for a single `deviceId`. In the `app_info_table`, one SPN maps to one `appInfoId` with b/459404842, one SPN will map to one `deviceId`.

Fix: 458001956
Bug: 440066697
Flag: com.android.healthfitness.flags.device_data_providers_api
Test: unit tests
Change-Id: I5ad2fa9e2292cda8fd411ecdc808123d7d82f635