Clear devices that are no longer advertised from device data sources
When a DDP omits a device it previously advertised, we should treat it
as no longer accessible via the DDP and we should e.g. exclude it from
matchmaking and from the set of device data sources we expose to
clients.

There may however still be data associated with the device that was
previously advertised, so we don't remove e.g. the app info entry or the
DDP metadata.

Fix: 460757169
Test: atest
Flag: com.android.healthfitness.flags.device_data_providers_api
Change-Id: Ic165414a74d3e289713e7af6cfdf58d37d79bce3