Check flag status before asserting non-nullness of DDP helper
It seems that this test was failing in environments where the DDP flag
was off due to the DDP helper being null and the requireNonNull check
failing. This non-nullness assertion should be done after checking the
flag status.

This issue also exists in getCurrentDeviceId, which has been patched in
this change too.

Fix: 461769963
Test: atest HealthFitnessUnitTests:com.android.server.healthconnect.HealthConnectServiceImplTest#getCurrentDeviceIdWithDisabledFlags_NullException_throwsUnsupportedError
Flag: EXEMPT BUGFIX
Change-Id: I9ce16b2b8f04c757668d2dae6be5f9d41707dec7