Add getDeviceDataSources API
This CL introduces the `getDeviceDataSources` API, which allows apps to retrieve a list of available device data sources. The API filters the results based on the caller's permissions. A device is included if the caller holds the READ permission for at least one of the data types supported by the device.

Fix: 440343237
Bug: 416236349
Test: atest DeviceDataSourceTest
Flag: com.android.healthfitness.flags.device_data_providers_api
Change-Id: I4c1d17485d4320bb935f8eda7f687ee005454a39