jarjar flags in controller tests
This ensures that the tests have their own copy of the flag definitions,
and avoids failures when the framework copy doesn't contain flags that
the tests depend on. This commonly occurs when the tests and installed
module have inconsistent versions, so the tests may still fail due to
changes in behaviour, but this at least avoids one common hurdle.

Bug: 460137985
Flag: EXEMPT TEST_ONLY
Test: atest CtsHealthConnectControllerTests
Change-Id: I494515e3468812aed66bba9361c975e92611fba8