Ignore implicit permissions in onboarding
When apps request BODY_SENSORS and target < SDK 36 this is translated
into an implicit request for READ_HEART_RATE. These can't be used to
access Health Connect, and so should be excluded.

This leaves usage stats unmodified for now, pending the decision on how
we handle updating the existing metric.

Bug: 461706432
Flag: EXEMPT BUGFIX
Test: atest HealthFitnessUnitTests
Change-Id: Id461e46975c74c60d4a3daef20b13db2466be13d