Create script to install the toolbox app as a privileged app
This will allow the toolbox app to call the DDP APIs. Since this
requires a privileged permission, it can't acquire this through normal
means. Instead, we install the toolbox app in the system apps directory,
and modify privapp-permissions to include the toolbox app. Credit to
jstembridge@ for the idea of simply writing a script to do this.

I've also added a button that allows advertising a device. I'll expand
this stuff later to include writing data etc.

Fix: 469677803
Test: n/a
Flag: EXEMPT TEST_ONLY

Change-Id: Iffa33b15919f3f38122376855aaf68655ad0b22c