Collect granted extra write permissions
We only care about the permissions that are granted, so there's no need
to pass around extra write permissions that aren't granted.

This also removes the use of a null or empty map to signify that the
extra permissions checks should be skipped, which is generally
non-obvious and so might lead to accidentally missed permission checks.

Bug: 460080622
Flag: EXEMPT PURE_REFACTOR
Test: atest HealthFitnessUnitTests
Change-Id: I54abc12f1b7106ff266d6a247fa06c9d31825fa9